requirements_test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements_test.in
#
--no-binary psycopg2

astroid==2.3.3            # via pylint
attrs==19.3.0             # via jsonschema, pytest
backcall==0.1.0           # via ipython
beautifulsoup4==4.6.0
boto3==1.11.8
botocore==1.14.8          # via boto3, s3transfer
certifi==2019.11.28       # via requests, sentry-sdk
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
codecov==2.0.15
coverage==5.0.3           # via codecov, django-coverage-plugin, pytest-cov
decorator==4.4.1          # via ipython, traitlets
directory-client-core==6.1.0  # via directory-sso-api-client
directory-components==35.10.0
directory-constants==20.10.0  # via directory-components
directory-sso-api-client==6.2.0
django-coverage-plugin==1.8.0
django-debug-toolbar==2.1
django-environ==0.4.5
django-extensions==2.2.6
django-filter==2.2.0
django-modelcluster==5.0.1  # via wagtail
django-pglocks==1.0.4
django-redis==4.10.0
django-staff-sso-client==1.0.1
django-storages[boto3]==1.8
django-taggit==1.2.0      # via wagtail
django-treebeard==4.3.1   # via wagtail
django==2.2.9
djangorestframework==3.9.4
docutils==0.15.2          # via botocore
draftjs-exporter==2.1.7   # via wagtail
entrypoints==0.3          # via flake8
factory-boy==2.12.0       # via wagtail-factories
faker==4.0.0              # via factory-boy
flake8==3.7.9
gunicorn==19.5.0
html5lib==1.0.1           # via wagtail
idna==2.8                 # via requests
importlib-metadata==1.4.0  # via jsonschema, pluggy, pytest
ipdb==0.12.3
ipython-genutils==0.2.0   # via traitlets
ipython==7.11.1           # via ipdb
isort==4.3.21             # via pylint
jedi==0.15.2              # via ipython
jmespath==0.9.4           # via boto3, botocore
jsonschema==3.2.0         # via directory-components
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via flake8, pylint
mohawk==0.3.4             # via sigauth
monotonic==1.5            # via directory-client-core
more-itertools==8.1.0     # via pytest, zipp
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.1           # via pytest, pytest-sugar
parso==0.5.2              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==6.2.2             # via wagtail
pip-tools==4.4.0
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.2     # via ipython
psycopg2==2.8.4
ptyprocess==0.6.0         # via pexpect
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pygments==2.5.2           # via ipython
pylint-django==2.0.13
pylint-plugin-utils==0.6  # via pylint-django
pylint==2.4.4             # via pylint-django, pylint-plugin-utils
pyparsing==2.4.6          # via packaging
pyrsistent==0.15.7        # via jsonschema
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-sugar==0.9.2
pytest==5.3.4
python-dateutil==2.8.1    # via botocore, faker
pytz==2019.3              # via django, django-modelcluster, wagtail
redis==3.3.11             # via django-redis
requests-mock==1.7.0
requests-oauthlib==1.3.0  # via django-staff-sso-client
requests==2.22.0          # via codecov, directory-client-core, requests-mock, requests-oauthlib, wagtail
s3transfer==0.3.2         # via boto3
sentry-sdk==0.13.5
sigauth==4.1.1            # via directory-client-core
six==1.14.0               # via astroid, django-coverage-plugin, django-extensions, django-pglocks, html5lib, jsonschema, mohawk, packaging, pip-tools, pyrsistent, python-dateutil, requests-mock, traitlets, w3lib, wagtail
sqlparse==0.3.0           # via django, django-debug-toolbar
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.3       # via faker
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via astroid
unidecode==1.1.1          # via wagtail
urllib3==1.25.8           # via botocore, requests, sentry-sdk
w3lib==1.21.0             # via directory-client-core
wagtail-factories==2.0.0
wagtail==2.7.1
wcwidth==0.1.8            # via prompt-toolkit, pytest
webencodings==0.5.1       # via html5lib
werkzeug==0.16.0
whitenoise==4.1.4
willow==1.3               # via wagtail
wrapt==1.11.2             # via astroid
zipp==2.0.1               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
